Changed the organization of the files.
Added packages, and changed how Zone and subclass constructors worked.

Now, there are 2 packages - one for Zones, and one for everything else - may add another one later.

Also, Zones that have subclasses all have a protected constructor now that takes a Color along with everything else.
This is the "real" constructor, that all other constructors should call, as it sets the color correctly without having to call any other methods.
This should NEVER be called externally, which is why Zones are in a seperate package.