Bots now make paths, and evaluate complete paths to determine which is best independant of base zone.
This involved a big overhaul of Bot as well as Survivor path. Also,
messages now have the ability to have Object attachments, as well as
message text (this may be removed later - it makes changing classes
and having messages still work MUCH easier, but you need to makes sure
immutable objects copy so that a change one Bot makes does not automatically
change every other Bot's copy - we're not allowing them to break the speed
limit of light).

Also, added a lot of good TODOs that should be implimented soon.